engine/requirements.txt
mcp>=1.0.0
python-telegram-bot>=20.7
openai>=1.30.0
eth-account>=0.11.0
pyyaml>=6.0.1
streamlit>=1.36.0
web3>=6.19.0
aiohttp>=3.9.0
requests>=2.31.0
